backend/requirements.txt
fastapi
uvicorn
pillow
pillow-heif
python-multipart
google-generativeai